test: Fix file not found error
Test updateTestFile was failing because it was expecting an output file
to exist from previous runs.

+++ src/test/java/com/udacity/jwdnd/course1/cloudstorage/model/SuperDuperFileModelTest.java
@@ -152,7 +152,9 @@ public class SuperDuperFileModelTest {
 
             // delete the output file from previous test runs.
             try {
-                Files.delete(Path.of(pathToOutputDoc));
+                if (Files.exists(Path.of(pathToOutputDoc))) {
+                    Files.delete(Path.of(pathToOutputDoc));
+                }
             } catch (Exception e) {
                 fail("Failed to delete existing file " + pathToOutputDoc + " : " + e.getMessage());
             }
